Added index and detail page route
