+++ www/deno.json
@@ -15,7 +15,7 @@
     "@std/fs": "jsr:@std/fs@0.224.0",
     "remark-gfm": "npm:remark-gfm@4.0.0",
     "remark-rehype": "npm:remark-rehype@11.1.1",
-    "rehype": "npm:unified@11.0.5",
+    "unified": "npm:unified@11.0.5",
     "rehype-prism-plus": "npm:rehype-prism-plus@2.0.0",
     "rehype-slug": "npm:rehype-slug@6.0.0",
     "rehype-autolink-headings": "npm:rehype-autolink-headings@7.1.1",
@@ -25,6 +25,7 @@
     "rehype-stringify": "npm:rehype-stringify@10.0.0",
     "remark-parse": "npm:remark-parse@11.0.0", 
     "hast-util-to-string": "npm:hast-util-to-string@3.0.0",
-    "@mdx-js/mdx": "npm:@mdx-js/mdx@3.0.1"
+    "@mdx-js/mdx": "npm:@mdx-js/mdx@3.0.1",
+    "vfile": "npm:vfile@6.0.3"
   }
 }
